+++ structurizr-core/src/com/structurizr/view/Routing.java
@@ -3,6 +3,7 @@ package com.structurizr.view;
 public enum Routing {
 
     Direct,
+    Curved,
     Orthogonal
 
 }